+++ x-pack/filebeat/cmd/agent.go
@@ -11,7 +11,6 @@ import (
 	"github.com/elastic/beats/v7/x-pack/libbeat/management"
 	"github.com/elastic/elastic-agent-client/v7/pkg/client"
 	"github.com/elastic/elastic-agent-client/v7/pkg/proto"
-	"github.com/elastic/elastic-agent-libs/mapstr"
 )
 
 func filebeatCfg(rawIn *proto.UnitExpectedConfig, agentInfo *client.AgentInfo) ([]*reload.ConfigWithMeta, error) {
@@ -35,23 +34,4 @@ func filebeatCfg(rawIn *proto.UnitExpectedConfig, agentInfo *client.AgentInfo) (
 	}
 
 	return configList, nil
-}
-
-func defaultProcessors() []mapstr.M {
-	// processors:
-	// - add_host_metadata:
-	// 	when.not.contains.tags: forwarded
-	// - add_cloud_metadata: ~
-	// - add_docker_metadata: ~
-	// - add_kubernetes_metadata: ~
-	return []mapstr.M{
-		{
-			"add_host_metadata": mapstr.M{
-				"when.not.contains.tags": "forwarded",
-			},
-		},
-		{"add_cloud_metadata": nil},
-		{"add_docker_metadata": nil},
-		{"add_kubernetes_metadata": nil},
-	}
 }
