+++ x-pack/filebeat/cmd/root.go
@@ -12,6 +12,7 @@ import (
 	"github.com/elastic/beats/v7/libbeat/processors"
 	"github.com/elastic/beats/v7/libbeat/publisher/processing"
 	"github.com/elastic/beats/v7/x-pack/libbeat/management"
+	"github.com/elastic/elastic-agent-libs/mapstr"
 
 	// Register the includes.
 	_ "github.com/elastic/beats/v7/x-pack/filebeat/include"
@@ -19,6 +20,7 @@ import (
 	_ "github.com/elastic/beats/v7/x-pack/libbeat/include"
 )
 
+// Name is the name of the beat
 const Name = fbcmd.Name
 
 // Filebeat build the beat root command for executing filebeat and it's subcommands.
@@ -33,4 +35,23 @@ func Filebeat() *cmd.BeatsRootCmd {
 	settings.ElasticLicensed = true
 	command := fbcmd.Filebeat(inputs.Init, settings)
 	return command
+}
+
+func defaultProcessors() []mapstr.M {
+	// processors:
+	// - add_host_metadata:
+	// 	when.not.contains.tags: forwarded
+	// - add_cloud_metadata: ~
+	// - add_docker_metadata: ~
+	// - add_kubernetes_metadata: ~
+	return []mapstr.M{
+		{
+			"add_host_metadata": mapstr.M{
+				"when.not.contains.tags": "forwarded",
+			},
+		},
+		{"add_cloud_metadata": nil},
+		{"add_docker_metadata": nil},
+		{"add_kubernetes_metadata": nil},
+	}
 }
